app/pyproject.toml
[[tool.uv.index]]
url = "https://pypi.tuna.tsinghua.edu.cn/simple"
default = true

[tool.aerich]
tortoise_orm = "config.TORTOISE_ORM"
location = "./migrations"
src_folder = "./."

[project]
name = "app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10,<3.13"
dependencies = [
    "aerich==0.7.2",
    "aiosqlite==0.17.0",
    "annotated-types==0.7.0",
    "anyio==4.6.2.post1",
    "async-timeout==4.0.3",
    "asyncpg==0.30.0",
    "certifi==2024.8.30",
    "click==8.1.7",
    "colorama==0.4.6",
    "dependency-injector==4.46.0",
    "dictdiffer==0.9.0",
    "dnspython==2.7.0",
    "email-validator==2.2.0",
    "exceptiongroup==1.2.2",
    "fastapi==0.115.5",
    "fastapi-cli==0.0.5",
    "gunicorn==23.0.0",
    "h11==0.14.0",
    "httpcore==1.0.7",
    "httptools==0.6.4",
    "httpx==0.27.2",
    "idna==3.10",
    "iso8601==1.1.0",
    "itsdangerous==2.2.0",
    "jinja2==3.1.4",
    "markdown-it-py==3.0.0",
    "markupsafe==3.0.2",
    "mdurl==0.1.2",
    "orjson==3.10.11",
    "packaging==24.2",
    "pillow==9.5.0",
    "pycryptodome==3.22.0",
    "pydantic==2.9.2",
    "pydantic-core==2.23.4",
    "pydantic-extra-types==2.10.0",
    "pydantic-settings==2.6.1",
    "pygments==2.18.0",
    "pyjwt==2.10.1",
    "pymysql==1.1.1",
    "pypika==0.48.9",
    "pypinyin==0.54.0",
    "python-dotenv==1.0.1",
    "python-multipart==0.0.17",
    "pytz==2024.2",
    "pyyaml==6.0.2",
    "redis==6.1.0",
    "rich==13.9.4",
    "ruff>=0.12.7",
    "shellingham==1.5.4",
    "sniffio==1.3.1",
    "starlette==0.41.2",
    "tomlkit==0.13.2",
    "tortoise-orm==0.20.1",
    "typer==0.13.0",
    "typing-extensions==4.12.2",
    "ujson==5.10.0",
    "uvicorn==0.32.0",
    "watchfiles==0.24.0",
    "websockets==14.1",
    # RAG相关依赖
    "langchain==0.3.7",
    "langchain-community==0.3.7",
    "langchain-openai==0.2.8",
    "sentence-transformers==3.1.1",
    "chromadb==0.5.5",
    "pypdf==5.0.1",
    "python-docx==1.1.2",
    "openpyxl==3.1.5",
    # 与 langchain==0.3.7 兼容的 numpy 版本（<2.0）。1.26.4 为最后一个 1.x 版本
    "numpy==1.26.4",
    "tiktoken==0.8.0",
]
